+++ src/main/java/newCar/event_page/model/dto/admin/AdminLoginDTO.java
@@ -1,0 +1,14 @@
+package newCar.event_page.model.dto.admin;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.Data;
+
+@Data
+public class AdminLoginDTO {
+
+    @NotEmpty
+    private String adminId;
+
+    @NotEmpty
+    private String password;
+}
