+++ src/main/java/newCar/event_page/model/entity/event/racing/PersonalityTest.java
@@ -3,7 +3,7 @@ package newCar.event_page.model.entity.event.racing;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
-import newCar.event_page.model.dto.PersonalityTestDTO;
+import newCar.event_page.model.dto.admin.AdminPersonalityTestDTO;
 import newCar.event_page.model.entity.TeamScore;
 
 @Entity
@@ -53,7 +53,7 @@ public class PersonalityTest {
     })
     private TeamScore choice2Scores;
 
-    public void update(PersonalityTestDTO personalityTestDTO)
+    public void update(AdminPersonalityTestDTO personalityTestDTO)
     {
         this.question = personalityTestDTO.getQuestion();
         this.choice1= personalityTestDTO.getChoices().get(0).getText();
